+++ bin/assembly_qc_report_943e0fb.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import json
+from tabulate import tabulate
 
 from report_modules.parsers.ncbi_fcs_adaptor_parser import parse_ncbi_fcs_adaptor_folder
 from report_modules.parsers.ncbi_fcs_gx_parser import parse_ncbi_fcs_gx_folder
@@ -20,11 +21,22 @@ from report_modules.parsers.kraken2_parser import parse_kraken2_folder
 from report_modules.parsers.hic_parser import parse_hic_folder
 from report_modules.parsers.circos_parser import parse_circos_folder
 from report_modules.report_printer import ReportPrinter
+from report_modules.parsers.params_parser import parse_params_json
 
 
 if __name__ == "__main__":
-    with open("constants.json", "r") as f:
-        constants_dict = json.load(f)
+    with open("params_json.json", "r") as f:
+        params_dict = json.load(f)
+        params_df = parse_params_json(params_dict)
+        params_table = tabulate(
+            params_df,
+            headers=["Parameter", "Value"],
+            tablefmt="html",
+            numalign="left",
+            showindex=False,
+        ).replace(
+            "<th>Parameter", '<th width="20%" style="text-align: left">Parameter', 1
+        )
 
     data_from_tools = {}
 
@@ -57,7 +69,8 @@ if __name__ == "__main__":
             "CIRCOS": "0.23-1",
             "MUMMER": "4.0.0",
         },
-        "CONSTANTS": constants_dict,
+        "PARAMS_DICT": params_dict,
+        "PARAMS_TABLE": params_table,
     }
 
     report_printer = ReportPrinter()